Updated camera follow script
Set the camera follow to behave in a way much more similar to Cave Story's
camera. Horizontally, the camera now rests much further ahead in the
direction the player is facing.

Also added a 'RoomData' script to the Level prefab. Simply do nothing with
it if the camera should behave normally. If you want the camera to be
stationary, set the bool to true, and set the x and y coordinate the
camera should sit at.

The vertical panning speed seems a bit slow. We may need to create
separate dampening variables for each of these. Try 0.5f for the vertical
dampening.